pyproject.toml
[tool.poetry]
name = "promptsearcher"
version = "0.1.0"
description = "PromptSearcher is an automatic tool designed to find the best prompt in both supervised and unsupervised scenarios. This project draws inspiration from traditional neural network learning techniques."
authors = ["octaviopavon <octavio.pavon@botman-ai.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.51.2"
polars = "^1.9.0"
numpy = "^2.1.2"
python-dotenv = "^1.0.1"
anthropic = "^0.36.1"
groq = "^0.11.0"
matplotlib = "^3.9.2"

[tool.poetry.dev-dependencies]
pytest = "^7.4.3"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
